fix: remove misleading 'rm' error message
Problem:
 When running a Session as a different user on posix systems you will
see a benign, but misleading, error message in the log that looks like:
```
rm: cannot remove '/tmp/openjd/session-name': Permission denied
```

This is caused by our two-phase cleanup of the Session working
directory. We start with a cross-user 'rm -rf' to delete everything in
the directory, and then do a same-user recursive delete. So, the working
directory is getting cleaned up but there is this error in logs that is
misleading.

Solution:
 The `rm -rf` just needs to be changed to `rm -rf <dir>/*` from `rm -rf
<dir>`.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>