feat: allow adding env vars when running an action
There are use cases where the entity running an action may wish to
define additional environment variables specifically for that action.
For example, an environment variable that contains an ID for a task that
is being run.

This adds an 'os_env_var' argument to the Session action run methods for
environment_enter, environment_exit, and run_task.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>