feat: modify logging to be easier to understand
The existing session logging is challenging to read in that it:
1. Is a continuous stream of log messages with no deliniation indicating
   which parts are relevant for which actions.
2. Intermixes the action output with logging from the library; which
   makes it challenging for the end-user to understand which bits are
   from which.

This commit does some cleanup of the logging to make it easier to read.
We now have banners indicating the start of running each action. We also
have subbanners indicating what phase of running the action we're in
(setup vs running).

As a bonus, I also added a new stdout/stderr handler that will be
specific to this implementation of Open Job Description; adding it as-is
to the specification would encode Python logging semantics into the
spec, and that's not desirable. This new stdout handler provides the
end-user a means by which they can dynamically modify the logger log
level of this library for the duration of the Session that is being run.
This is added as a debugging mechanism so that users can see the
contents of files that are materialized to disk, and any other verbose
output as added in future. The allowable forms of this handler are:
openjd_session_runtime_loglevel: ERROR
openjd_session_runtime_loglevel: WARNING
openjd_session_runtime_loglevel: INFO
openjd_session_runtime_loglevel: DEBUG

i.e. The values of the prefix correspond to Python logger levels.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>