feat!: reuse ParameterValueType from model package
This Sessions package defines its own 'ParameterType' for the type of
Job & Task Parameters. The 'openjd-model' package also defines a
ParameterValueType that is identical. It's awkward to have to translate
enums between the two libraries; they should just be using the same one.

This deletes ParameterType from this package in favour of using
ParameterValueType from 'openjd.model'.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>