feat!: unify parameter data shapes with openjd-model
Problem:
 The openjd-model package defines parameters using a dictionary/map from
parameter name to parameter attributes (type + value). It does this
because the primary use-case there is being able to quickly find a
specific parameter from a parameter set/list. In this package we defined
the parameters using a list of parameters where each of those list
elements contained the name, type, and value of the parameter. This
non-compatibility leads to a need to reshape a parameter list just to
pass values from the openjd-model methods to the Session.

Solution:
 Unify the job and task parameter shapes in this package to match the
format in the openjd-model package. We now accept dictionary/mapping
of parameter name to attributes.

**BREAKING CHANGE** - The argument types to pass job/task parameters to
the Sessions constructor and run_task() method has changed. Please read
the docstring for the new type information.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>