chore: Use win apis to run process (#58)
Problem:

The Windows RunAs (aka Impersonation) and Notify features were being managed via powershell scripts.
a) this involved passing a username and password via command line.
b) it was specifcally using powershell 7 (pwsh.exe), which is not default on Windows.
c) the notify performance was slower

Solution:

Remove this powershell layer and use Win apis to start and notify the process.

We use python subprocess.Popen for the non-impersonation case. For impersonation, we derive from Popen and run CreateProcessWithLogonW for our use case. Notify, for impersonation, is now a python script running win apis.

We still use one powershell script to set the env vars and run the command. Here we switch it back to running with powershell 5 rather than 7.

---------

Signed-off-by: Authement <maauth@amazon.com>