chore: enable cross-user linux tests in CI
Summary:
This library has functionality that allows it to run subprocesses as a user
that is different from the current user. We have unit/integration tests that
verify this functionality is working. Right now, a developer is expected to run
the scripts/run_sudo_tests.sh set of tests before putting up a pull request
to make sure that a change doesn't break the functionality. This isn't reliable
so we should have these tests enabled in the GitHub CI; this does that.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>