feat!: remove group property from WindowsSessionUser
Summary:

The design of the WindowsSessionUser is currently heavily influenced by
the posix design in that it requires a Group for setting directory ACLs
of the Session Working Directory. Windows has finer grained ACLs for
controlling access to files and directories, so we should be exposing an
interface that works better with that instead.

Solution:

We no longer accept a group in the WindowsSessionUser. The Session
Working Directory's ACLS are now set to:
1. Process owner -- full control
2. Given user -- modify/read/write/execute

**BREAKING CHANGE**
The "group" property of WindowsSessionUser has been removed.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>