feat: resolve windows command location prior to run (#116)
Problem:

When running a subprocess on Windows without a shell, as we do, there is
no guarantee that Windows will find the and run the first instance of
the command executable on the user's PATH. We've seen this manifest as a
user having mingw bash installed and first in their PATH, but running an
action with command "bash" ends up finding the WSL bash executable
instead. To ensure that the correct command is run you must provide the
full path to the command.

Solution:

It's multifacted. The main factor is that the user that's running the
main process may not be the same user that an Action is running as, and
thus may have different default PATH and may not even be able to
access/list the same files/directories as the Action user. So, we:

1. If running the Action as the same user as the main process, then we
   use shutil.which() to resolve the command's full path, and then run
   that. Taking care to prepend the Session Working Dir to the PATH if
   an openjd_env has previously been done that sets the PATH.

2. If running the Action as a different user, then we run "cmd /C 'which
   <command name>'" and parse the output to figure out where the command
   is located. We run this command with the same env vars and working
   directory as the Action will be run with.

There are limitations of this approach, but it seems to be the best that
I can currently figure out. Notably, if a command is a relative path
then 'where' cannot resolve the full command path, so we just have to
pass the command through as relative and hope for the best.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>