fix: eliminate TranslateName usage on Windows systems
Problem:

A customer has reported that a call to the TranslateName Win32 API is
failing with an access denied error when constructing a
WindowsSessionUser with their AD setup; their AD does not support
SAM account names.

Solution:

The TranslateName call was not necessary, so it has been removed. It
appears to have been put in place to be able to extract a domain for a
LogonUser call that is used to check that the user+password are correct.
However, LogonUser does not require that a domain be passed to it.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>