feat: Add run_subprocess function to the Session class
This new function provides the ability to run an ad hoc command within
the session environment, choosing whether to use the environment
variables from the session enter actions or to ignore them. The
parameters for the command are direct, so you don't have to construct a
temporary Action or StepAction object to run a simple command.

Signed-off-by: Mark <399551+mwiebe@users.noreply.github.com>